Refactored locks in CDC driver.
Global lock during read or write operations replaced with locking for
queue operations. Note that in this case size reading or peeking of
the last element do not require locking, because queue content is not
modified (single producer - single consumer). The latency in processing
of other ISR's is reduced. CDC throughput is slightly decreased too,
but only without optimizations, because transmission begins during write
operation and in some cases might be finished before a next chunk
of data will be sent, so packets will not be queued.