Refactored architecture of USB stack.
Device and configuration descriptors are post-processed in the USB Control
module. This update simplifies descriptor functors in driver classes and
reduces code duplication in UsbDevice classes.

Added base class for USB endpoints to provide the possibility of different
endpoint implementations for SIE and DMA endpoints on LPC17xx.

Functions usbDevGetParameter and usbDevSetParameter are replaced with
usbDevGetSpeed and usbDevSetPower. Other functions became unnecessary
because of the direct access to parameters from descriptor
post-processing functions.

UsbDevice must be switched to Connected state from user code with
the usbDevSetConnected function.

Added minor fixes in configuration structures of HID classes.